feat: Change implementation to add locales Also - Replace server implementation to add a rollup configuration to embed all code and avoid API calls
