+++ chrome-extension/manifest.json
