+++ chrome-extension/popup.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html>
+<body>
+<label for="lang">Language</label>
+<select id="lang">
+    <option value="en">English</option>
+    <option value="fr">French</option>
+</select>
+<textarea id="input" rows="6" cols="30" placeholder="E.g. dinner friday 7pm"></textarea>
+<br>
+<button type="button" id="add">Add</button>
+<div id="error" style="color: red"></div>
+
+<script src="popup.js"></script>
+</body>
+</html>
