--- chrome/popup.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <body>
-    <textarea id="input" rows="6" cols="30" placeholder="E.g. dinner friday 7pm"></textarea>
-    <br>
-    <button type="button" id="add">Add</button>
-    <div id="error" style="color: red"></div>
-
-    <script src="popup.js"></script>
-  </body>
-</html>
