--- package.json
@@ -1,12 +1,0 @@
-{
-  "dependencies": {
-    "@fastify/formbody": "^7.0.0",
-    "chrono-node": "^2.7.4",
-    "dayjs": "^1.11.10",
-    "fastify": "^4.0.0-rc.2"
-  },
-  "type": "module",
-  "scripts": {
-    "start": "node server.js"
-  }
-}
