+++ rollup.config.mjs
@@ -1,0 +1,11 @@
+import resolve from 'rollup-plugin-node-resolve';
+import commonJS from 'rollup-plugin-commonjs'
+
+export default {
+  plugins: [
+    resolve(),
+    commonJS({
+      include: ['node_modules/**',]
+    }),
+  ],
+};
