+++ src/helpers.js
@@ -1,5 +1,7 @@
-import * as chrono from 'chrono-node';
+import {en, fr} from 'chrono-node';
 import dayjs from 'dayjs';
+
+const locales = {en, fr}
 
 function dateRange(start, end, isAllDay) {
   let formatString = 'YYYYMMDD[T]HHmmss';
@@ -11,12 +13,15 @@ function dateRange(start, end, isAllDay) {
   return [start, end].map(t => t.format(formatString)).join('/');
 }
 
-function parse(text) {
+function parse(text, lang = "en") {
+  console.log("lang", lang)
   if (!text) {
     throw new Error('invalid input text');
   }
 
-  const results = chrono.parse(text);
+
+  const locale = locales[lang] || en
+  const results = locale.parse(text)
 
   if (results.length === 0) {
     throw new Error('could not find time data');
@@ -47,13 +52,13 @@ function parse(text) {
   }
 
   const dates = dateRange(start, end, isAllDay);
-  return { text: eventTitle, dates }
+  return {text: eventTitle, dates}
 }
 
-export function createEventUrl(text) {
+export function createEventUrl(text, lang) {
   let data;
   try {
-    data = parse(text);
+    data = parse(text, lang);
     console.log(data);
   } catch (err) {
     console.log(err);
@@ -66,13 +71,3 @@ export function createEventUrl(text) {
   const params = new URLSearchParams(data);
   return `${baseUrl}?${params}`;
 }
-
-function main() {
-  const input = "Brunch with Gary and Mira tomorrow at 10:30am";
-  // const input = "Event now";
-
-  const url = createEventUrl(input);
-  console.log(url);
-}
-
-// main();
